+++ NEWS.md
@@ -3,6 +3,7 @@ CARTO Mobile SDK 4.3.1
 
 ### Changes/fixes:
 
+* Fixed a crashing issue with ClusteredVectorLayer
 * Minor documentation updates